requirements.txt
streamlit>=1.18
google-genai>=0.0.1
python-dotenv>=1.0
